+++ bflib/src/menu/mod.rs
@@ -198,9 +198,14 @@ pub(super) fn init_for_slot(ctx: &Context, lua: MizLua, slot: &SlotId) -> Result
     let mc = MissionCommands::singleton(lua)?;
     let add_jtac = |side, gid, ucid| -> Result<()> {
         let mut roots: FxHashMap<String, GroupSubMenu> = FxHashMap::default();
-        let root = mc.add_submenu_for_group(gid, "JTAC".into(), None)?;
+        let mut root = mc.add_submenu_for_group(gid, "JTAC".into(), None)?;
+        let mut n = 0;
         for jtac in ctx.jtac.jtacs() {
             if jtac.side == side {
+                if n >= 9 {
+                    root = mc.add_submenu_for_group(gid, "NEXT>>".into(), Some(root.clone()))?;
+                    n = 0;
+                }
                 let near = ctx
                     .db
                     .objective(&jtac.location.oid)
@@ -219,7 +224,8 @@ pub(super) fn init_for_slot(ctx: &Context, lua: MizLua, slot: &SlotId) -> Result
                         e.insert(root).clone()
                     }
                 };
-                jtac::add_menu_for_jtac(&ctx.db, side, root, lua, gid, jtac, ucid).context("adding jtac menu")?
+                jtac::add_menu_for_jtac(&ctx.db, side, root, lua, gid, jtac, ucid)
+                    .context("adding jtac menu")?
             }
         }
         Ok(())